fix #2

+++ cogs/cb.py
@@ -137,10 +137,10 @@ class CB(commands.Cog):
                     newRound.messageId = message.id
                     self.queue.rounds.append(newRound)
 
-                if self.queue.currentRound <= 45:
-                    # remove oldest round embed
-                    message = await self.client.get_channel(self.activeChannel).fetch_message(
-                        self.queue.rounds[0].messageId)
+                # remove oldest round embed
+                message = await self.client.get_channel(self.activeChannel).fetch_message(
+                    self.queue.rounds[0].messageId)
+                if message:
                     await message.delete()
                     self.queue.rounds.pop(0)
 
@@ -171,21 +171,23 @@ class CB(commands.Cog):
                 self.queue.currentBoss = 1
                 await ctx.send(f'Proceeding to round {self.queue.currentRound}.')
 
-                # add next round
-                newRound = Round()
-                newEmbed = makeQueueEmbed(newRound, self.queue.currentRound + 2)
-                channel = self.client.get_channel(self.activeChannel)
-                message = await channel.send(embed=newEmbed)
-                for emoji in self.emojis:
-                    await message.add_reaction(emoji)
-                newRound.messageId = message.id
-                self.queue.rounds.append(newRound)
+                if self.queue.currentRound <= 43:
+                    # add next round
+                    newRound = Round()
+                    newEmbed = makeQueueEmbed(newRound, self.queue.currentRound + 2)
+                    channel = self.client.get_channel(self.activeChannel)
+                    message = await channel.send(embed=newEmbed)
+                    for emoji in self.emojis:
+                        await message.add_reaction(emoji)
+                    newRound.messageId = message.id
+                    self.queue.rounds.append(newRound)
 
                 # remove oldest round embed
                 message = await self.client.get_channel(self.activeChannel).fetch_message(
                     self.queue.rounds[0].messageId)
-                await message.delete()
-                self.queue.rounds.pop(0)
+                if message:
+                    await message.delete()
+                    self.queue.rounds.pop(0)
             else:
                 self.queue.currentRound = round
                 self.queue.currentBoss = 1
